keikokunakusimasita

+++ Screen.cpp
@@ -35,14 +35,14 @@ void Screen::Shake(int minX, int maxX, int minY, int maxY, bool is) {
 
 
 
-void Screen::DrawBox(int x, int y, int w, int h, float angle, unsigned int color, FillMode fillMode) {
+void Screen::DrawBox(float x, float y, float w, float h, float angle, unsigned int color, FillMode fillMode) {
 	Vec2 tmp(x - Scroll.x, y - Scroll.y);
 	x = tmp.x * Zoom.x + Worldcenter.x + ScreenShake.x; 
 	y = tmp.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
-	Novice::DrawBox(x, y, w, h, angle, color, fillMode);
+	Novice::DrawBox(static_cast<int>(x), static_cast<int>(y), w, h, angle, color, fillMode);
 }
 
-void Screen::DrawTriangle(int x1, int y1, int x2, int y2, int x3, int y3, unsigned int color, FillMode fillMode) {
+void Screen::DrawTriangle(float x1, float y1, float x2, float y2, float x3, float y3, unsigned int color, FillMode fillMode) {
 	Vec2 tmp1(x1 - Scroll.x, y1 - Scroll.y);
 	Vec2 tmp2(x2 - Scroll.x, y2 - Scroll.y);
 	Vec2 tmp3(x3 - Scroll.x, y3 - Scroll.y);
@@ -52,41 +52,41 @@ void Screen::DrawTriangle(int x1, int y1, int x2, int y2, int x3, int y3, unsign
 	y2 = tmp2.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
 	x3 = tmp3.x * Zoom.x + Worldcenter.x + ScreenShake.x; 
 	y3 = tmp3.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
-	Novice::DrawTriangle(x1, y1, x2, y2, x3, y3, color, fillMode);
+	Novice::DrawTriangle(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2), static_cast<int>(y2), static_cast<int>(x3), static_cast<int>(y3), color, fillMode);
 }
 
-void Screen::DrawLine(int x1, int y1, int x2, int y2, unsigned int color) {
+void Screen::DrawLine(float x1, float y1, float x2, float y2, unsigned int color) {
 	Vec2 tmp1(x1 - Scroll.x, y1 - Scroll.y);
 	Vec2 tmp2(x2 - Scroll.x, y2 - Scroll.y);
 	x1 = tmp1.x * Zoom.x + Worldcenter.x + ScreenShake.x; 
 	y1 = tmp1.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
 	x2 = tmp2.x * Zoom.x + Worldcenter.x + ScreenShake.x; 
 	y2 = tmp2.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
-	Novice::DrawLine(x1, y1, x2, y2, color);
+	Novice::DrawLine(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2), static_cast<int>(y2), color);
 }
 
-void Screen::DrawEllipse(int x, int y, int radiusX, int radiusY, float angle, unsigned int color, FillMode fillMode) {
+void Screen::DrawEllipse(float x, float y, float radiusX, float radiusY, float angle, unsigned int color, FillMode fillMode) {
 	Vec2 tmp(x - Scroll.x, y - Scroll.y);
 	x = tmp.x * Zoom.x + Worldcenter.x + ScreenShake.x; 
 	y = tmp.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
-	Novice::DrawEllipse(x, y,  radiusX * Zoom.x,  radiusY * Zoom.y,  angle,  color, fillMode);
+	Novice::DrawEllipse(static_cast<int>(x), static_cast<int>(y),  radiusX * Zoom.x,  radiusY * Zoom.y,  angle,  color, fillMode);
 }
 
-void Screen::DrawSprite(int x, int y, int textureHandle, float scaleX, float scaleY, float angle, unsigned int color) {
+void Screen::DrawSprite(int x, int y, int textureHandle, int scaleX, int scaleY, int angle, unsigned int color) {
 	Novice::DrawSprite( 
 		x + Worldcenter.x + ScreenShake.x, 
 		y * -1 + Worldcenter.y - ScreenShake.y,  
 		textureHandle,  scaleX,  scaleY,  angle,  color);
 }
 
-void Screen::DrawSpriteRect(int destX, int destY, int srcX, int srcY, int srcW, int srcH, int textureHandle, float scaleX, float scaleY, float angle, unsigned int color) {
+void Screen::DrawSpriteRect(int destX, int destY, int srcX, int srcY, int srcW, int srcH, int textureHandle, int scaleX, int scaleY, int angle, unsigned int color) {
 	Novice::DrawSpriteRect(
 		destX + Worldcenter.x + ScreenShake.x,
 		destY * -1 + Worldcenter.y - ScreenShake.y,
 		srcX, srcY, srcW, srcH, textureHandle, scaleX, scaleY, angle, color);
 }
 
-void Screen::DrawQuad(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4, int srcX, int srcY, int srcW, int srcH, int textureHandle, unsigned int color) {
+void Screen::DrawQuad(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4, float srcX, float srcY, float srcW, float srcH, float textureHandle, unsigned int color) {
 	Vec2 tmp1(x1 - Scroll.x, y1 - Scroll.y);
 	Vec2 tmp2(x2 - Scroll.x, y2 - Scroll.y);
 	Vec2 tmp3(x3 - Scroll.x, y3 - Scroll.y);
@@ -99,7 +99,7 @@ void Screen::DrawQuad(int x1, int y1, int x2, int y2, int x3, int y3, int x4, in
 	y3 = tmp3.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
 	x4 = tmp4.x * Zoom.x + Worldcenter.x + ScreenShake.x; 
 	y4 = tmp4.y * Zoom.y * -1 + Worldcenter.y - ScreenShake.y;
-	Novice::DrawQuad(x1, y1, x2, y2, x3, y3, x4, y4, srcX, srcY, srcW, srcH, textureHandle, color);
+	Novice::DrawQuad(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2), static_cast<int>(y2), static_cast<int>(x3), static_cast<int>(y3), static_cast<int>(x4), static_cast<int>(y4), srcX, srcY, srcW, srcH, textureHandle, color);
 }
 
 void Screen::DrawQuad2(int x1, int y1, int quadw, int quadh, int srcX, int srcY, int srcW, int srcH, int textureHandle, unsigned int color) {
